Create setup1.ps1
Created w/new name so as not to impact this and other DP-3XXX courses that currently use the 'setup.ps1' at root of repo.

Changes in this new file consist of a proposed alteration to the script logic such that it allows a variable to be input when running the .ps1 for $suffix, i.e:

./mslearn-databricks/setup.ps1 "<variable desired>"

If provided, it would allow an ALH to submit a known value (such as a lab instanceID) or other value that allows ACPs to match a specific ResourceGroup name, and also a specific name for the Databricks workspace.  

If no value is provided, the script would function as befure using a 7-char alphanumeric value.